--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,7 +1,8 @@
 [workspace]
 resolver = "2"
 members = [
-    "yake_rust"
+    "yake_rust",
+    "python",
 ]
 
 [profile.dev]