--- a/python/src/lib.rs
+++ b/python/src/lib.rs
@@ -55,7 +55,7 @@ impl Yake {
     ///
     /// Get the n best keywords from text.
     #[pyo3(signature = (text, *, n))]
-    pub fn get_n_best(&mut self, py: Python, text: String, n: usize) -> Vec<(String, f64)> {
+    pub fn get_n_best(&self, py: Python, text: String, n: usize) -> Vec<(String, f64)> {
         py.allow_threads(|| get_n_best_sequential(n, text, &self._stopwords, &self._config))
     }
 }