--- a/python/src/yake_rust/__init__.py
+++ b/python/src/yake_rust/__init__.py
@@ -1,0 +1,8 @@
+"""Python interface to yake-rust."""
+
+from ._interface import Yake, YakeConfig
+
+__all__ = (
+    "Yake",
+    "YakeConfig",
+)