--- a/python/tox.ini
+++ b/python/tox.ini
@@ -1,0 +1,14 @@
+[tox]
+envlist = py39,py310,py311,py312,py313
+download = true
+minversion = 4.0.0
+isolated_build = True
+
+[testenv]
+deps =
+    -rrequirements-dev.txt
+commands =
+    ruff check
+    ruff format --check --diff
+    mypy --cache-dir="{envdir}/mypy_cache"
+    pytest