Open issues:

**RuntimeError** (Edit: Fixed)
The concurrency test has some warning:
```
PytestUnhandledThreadExceptionWarning: Exception in thread Thread-4 (get_n_best)

  Traceback (most recent call last):
    File "<...>threading.py", line 1052, in _bootstrap_inner
      self.run()
    File "<...>threading.py", line 989, in run
      self._target(*self._args, **self._kwargs)
  RuntimeError: Already borrowed
```
I suppose it is even possible that right now the test is only passing because the threads just die immediately and thus "finish quickly". I should change the test to check for exceptions in the threads. This issue should then be investigated. It is imperative that we have concurrency in my opinion. Shouldn't multiple threads be able to each have an immutable reference to the same data and do whatever? Are we missing some send/sync stuff?

**Typing** (Edit: Fixed)
`Yake.__init__` takes either `language` or `stopwords`, but not both, and it must have one of them. I tried to represent this in the .pyi file with `typing.overload` but did not succeed. It would be nice if we could get this working.

**Building and environments**
Maturin can build sdists and wheels - I think - I would like to make sure the sdist is fine, and I would like to look into setting up some github workflows for building wheels for different systems.

**Readme and license** (Edit: Fixed)
I can reference the outer `README.md` and `LICENSE` files with a relative path in `python/Cargo.toml`, but then tox won't work because it won't find those files in the isolated build environment (`MANIFEST.in` cannot fix this because it cannot specify paths in its own parent directory). I'd rather have those in place and just drop tox (because we ship the package, not the tests/linting), but I am concerned about sdist distributions maybe having some issue with this as well, so I would at least like to test sdist stuff more.

**Version** (Edit: Resolved)
Do we want the python package to have the same version as the rust crate? If so, we would have to keep `version` in `python/Cargo.toml` aligned with `yake_rust/Cargo.toml`. We could of course use separate versions for them. If we want the same, it would be nice if we did not have to do that manually.

**Uploading** (Edit: Fixed)
I have not done anything with regards to uploading anything to pypi. I was thinking I would set up permissions for this project once the building is completely in place.