@bunny-therapist 2025-01-25
Alright. Fixed the concurrency issue. I just needed to make the self reference immutable.