@bunny-therapist 2025-01-25
I suppose the python package should have its own README with python code examples.