@xamgore 2025-01-25
Fantastic work 👀

Could you extend CONTRIBUTING.md? In case anyone opens an issue, we might ask him to propose a PR as well.

I have looked through the code, looks fine. I would switch off a few clippy rules.

- Does pyo3 copies data when passing it through?
- Curious, why yake-rust is not so fast on small texts and where is the problem
- Is it okay if I replace std::collections::HashMap with FxHashMap or AHashMap, which are not cryptographically safe and therefore faster?
- I am in the middle of rewriting the README, is it ok if we postpone the release till next Friday?