@bunny-therapist 2025-01-25
> Fantastic work 👀
> 
> Could you extend CONTRIBUTING.md? In case anyone opens an issue, we might ask him to propose a PR as well.
>

Sorry, I don't quite understand what you are asking me to add.

> I have looked through the code, looks fine. I would switch off a few clippy rules.
> 
>     * Does pyo3 copies data when passing it through?
>

My understanding is that it converts python objects to rust objects and vice versa, based on some rules. So `str` becomes `String` going in and the returned `Vec` becomes a `list`. Not sure exactly what it means to "copy data" here. The original python object still exists in python, but you now have a converted rust object as well, so presumably "yes"?

>     * Curious, why yake-rust is not so fast on small texts and where is the problem
> 

I was a bit surprised by this too, because when I was first writing python bindings for yake-rust (before you and I got involved in the crate), this was not the case. I suspect we changed something that made it slower. Did you compare segtok with unicode-segmentation performance?

Of course, going between rust and python will incur some performance cost. There are some options one can try to improve it; but I don't want to do something I don't understand. Adding "frozen" to the Yake pyclass actually improved performance significantly.

>     * Is it okay if I replace std::collections::HashMap with FxHashMap or AHashMap, which are not cryptographically safe and therefore faster?
> 

You mean for stopwords and punctuation? As long as it is possible to convert a regular hashmap into these I think it is fine. Stopwords and punctuation will come into the rust code from python as `HashSet`, so I would have to convert it before putting it in the config. However, that is only done at class instantiation anyway (I actually decided to go with a class for this very purpose - I wanted to make sure we could reuse rust objects like stopwords etc without passing them back and forth between python and rust).

>     * I am in the middle of rewriting the README, is it ok if we postpone the release till next Friday?
> 

Sure.

> 
> I'll read the PR once again, as this time I've done it with my phone.

It is still just a draft. I need to understand the actual building process better and look into some stuff. Though we can of course merge this before everything is 100% done. I will convert it from a draft to a real PR when I feel good about it.