@bunny-therapist 2025-01-26
I think we will just go with having different versioning for the python package and the rust crate. Could be useful, and I can't think of a way to unify them that isn't ridiculous.