@bunny-therapist 2025-01-26
I don't quite understand this or if this is something that we should do or not: https://pyo3.rs/v0.23.4/performance.html#disable-the-global-reference-pool