@xamgore 2025-01-27
Well, replacing `segtok` with just `.split(".").split_whitespace()` gives a 1000x speed up. When running a test with a small text (180 bytes) the time is mostly spent on initializing lazy regex automatons, but further tests run faster: 6ms vs. 100μs. In my benchmarks I see the slowest time, the median, the mean, the fastest.