@bunny-therapist 2025-01-27
Interesting. Do you know if this also applies when calling from Python? I would assume so, but I will check.