@bunny-therapist 2025-01-27
I don't actually see that kind of difference when running from python. For the short text, the time is
```
0.0001104000
0.0001050000
0.0001055000
0.0001173000
0.0001231000
0.0000935000
0.0001338000
0.0001003999
0.0001002001
0.0000936999
0.0000965000
```
Though I guess they are all like 100 microseconds like you say.