@bunny-therapist 2025-01-27
For comparion, this is LIAAD/yake for the same short text:
```
0.0004106000
0.0003625000
0.0002954000
0.0004134000
0.0005442001
0.0003265999
0.0002978999
0.0003141001
0.0002775000
0.0003502000
0.0002877000
```