@bunny-therapist 2025-01-27
Looking at the above figures, we _are_ faster than LIAAD/yake. The reason the tests are flaky and we do not always beat LIAAD/yake might be due to outliers. Here is a failing test:

the assertion is `assert dt_rust < dt_liaad`
and it evaluates to
```
assert 0.010449799941852689 < 0.0009973000269383192
```
So in this case, yake-rust was like a 100 times slower.