@bunny-therapist 2025-01-27
Oh, ok. I think the "warm up" absolutely matters. I removed the "flaky" decorator and just ran the test for the short text, and it failed to beat LIAAD/yake every time. I added an extra run of yake-rust at the start of the test, discarding the result (just to have called it once already), and now yake-rust beats LIAAD/yake consistently.

I think I was just confused by "flaky" and the order of tests or something. I think the first time we call `get_n_best` it is slow, then faster subsequent times, beating LIAAD/yake consistently.

So it really _is_ a "warm-up" issue.

Would it be worth it to deal with this somehow? Or is it just acceptable?