@xamgore 2025-01-28
> Does it help segtok that regexs are lazy?

I'll come back later to answer this.

For both benchmarks and latency-critical applications, I can expose a function like `warm_up` or `init`.