@bunny-therapist 2025-01-28
I have added a github workflow triggered on "release" (so creating release via GitHub interface) which will build and upload to test.pypi (I will replace with the real pypi once it is confirmed to work).

The python wheels are platform dependent, so they need to be built for multiple platforms. I am also building for every supported python version. That could be dropped if we opt-in to the "abi3" interface, but it will allegedly come with an unknown performance cost. (I also have not really figured out how to do it properly.)