@bunny-therapist 2025-01-28
I added CI that runs the python tests on pushes and PRs to master.

Hopefully we won't be using up all of Kyle's free github runner minutes. Apparently there are 2000 minutes per year, and the CI seems to consume like 5, so 400 runs, then the release jobs is probably more than that.

Should be fine, but hopefully we won't push just one commit at a time. We could also adjust this. It's just that if I am going to publish from GitHub master branch, there should be checks that verify that it is working. We also have no git hooks that check this.