@bunny-therapist 2025-01-28
Note that the python jobs must necessarily build the rust crate also. I suppose we could have e.g. an initial CI job that runs "cargo check" or something that the other jobs depend on, so those jobs won't run if the code is fubar (of course, git hooks make this less likely).