@xamgore 2025-01-28
> Apparently there are 2000 minutes per year

We may ask Github team for more time.