@bunny-therapist 2025-01-29
@xamgore When you publish yake-rust crate, do you do that from the yake_rust directory? Is that how it works? I was just concerned about the project root Cargo.toml containing "python" as a member when we don't want to publish what is in there. (I have no experience with publishing crates.)