@xamgore 2025-01-29
https://doc.rust-lang.org/cargo/reference/manifest.html#the-publish-field