@bunny-therapist 2025-01-29
Ah, so I don't need the version field in Cargo.toml. I will replace it with publish = false then.