@xamgore 2025-02-04
Do you wanna bring another 7-10% speedup for free?

before:
```
Timer precision: 41 ns
bench          fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ text_3kb    3.978 ms      │ 13.9 ms       │ 4.048 ms      │ 4.077 ms      │ 2443    │ 2443
├─ text_100kb  139 ms        │ 145.6 ms      │ 140.2 ms      │ 140.5 ms      │ 100     │ 100
╰─ text_170b   187.8 µs      │ 401.5 µs      │ 191.7 µs      │ 193.4 µs      │ 51673   │ 51673
```

after:
```
Timer precision: 41 ns
bench          fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ text_3kb    3.808 ms      │ 13.95 ms      │ 3.86 ms       │ 3.879 ms      │ 2568    │ 2568
├─ text_100kb  129.9 ms      │ 136.1 ms      │ 131.2 ms      │ 131.5 ms      │ 100     │ 100
╰─ text_170b   175.4 µs      │ 370.3 µs      │ 178.4 µs      │ 179.8 µs      │ 55580   │ 55580
```